Changed app structure.
