--- Rolique/Model/Model.h
@@ -1,19 +1,0 @@
-//
-//  Model.h
-//  Model
-//
-//  Created by Bohdan Savych on 8/5/19.
-//  Copyright © 2019 Rolique. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for Model.
-FOUNDATION_EXPORT double ModelVersionNumber;
-
-//! Project version string for Model.
-FOUNDATION_EXPORT const unsigned char ModelVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Model/PublicHeader.h>
-
-
