+++ Rolique/Model/Models/Actions/Action.swift
@@ -7,8 +7,7 @@
 //
 
 import Foundation
-import Network
-import Env
+import Networking
 
 public class Action: Codable {
   let type, sender, test: String
