+++ Rolique/RemoteTodayExtension/RemoteViewController.swift
@@ -9,19 +9,9 @@
 import UIKit
 import NotificationCenter
 import UsersWidget
-import Model
 import Utils
 
-// MARK: - Userable
-extension User: Userable {
-  public var name: String {
-    return self.slackProfile.realName
-  }
-  
-  public var thumbnailURL: URL? {
-    return URL(string: (self.slackProfile.image48 ?? self.slackProfile.image32 ?? ""))
-  }
-}
+
 
 final class RemoteViewController: UsersViewController {
   private let userManager: UserManager = UserManagerImpl()
