--- Rolique/UI/UI.h
@@ -1,19 +1,0 @@
-//
-//  UI.h
-//  UI
-//
-//  Created by Bohdan Savych on 8/5/19.
-//  Copyright © 2019 Bohdan Savych. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for UI.
-FOUNDATION_EXPORT double UIVersionNumber;
-
-//! Project version string for UI.
-FOUNDATION_EXPORT const unsigned char UIVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <UI/PublicHeader.h>
-
-
