+++ Rolique/VacationTodayExtension/VacationViewController.swift
@@ -10,18 +10,7 @@ import UIKit
 import NotificationCenter
 import UsersWidget
 import Utils
-import Model
 
-// MARK: - Userable
-extension User: Userable {
-  public var name: String {
-    return self.slackProfile.realName
-  }
-  
-  public var thumbnailURL: URL? {
-    return URL(string: (self.slackProfile.image48 ?? self.slackProfile.image32 ?? ""))
-  }
-}
 
 class VacationViewController: UsersViewController {
   private let userManager: UserManager = UserManagerImpl()
